Ajout de test unitaire

+++ model/src/main/java/DAO/DataDAO.java
@@ -15,11 +15,11 @@ public class DataDAO extends AbstractDAO{
 	private final static String sqlGetMapByID ="{call insertdata(?,?)}";
 
 
-	public void save(int player , int time){
+	public void save(int player , float time){
 		final CallableStatement callStatement = prepareCall(sqlGetMapByID);
 	    try {
 			callStatement.setInt(1, player);
-			callStatement.setInt(2, time);
+			callStatement.setFloat(2, time);
 			callStatement.execute();
